make verified content completely read-only
The only non-corrupting modification to content stored in cacache is to
delete it, and that requires write permission on the containing
directory rather than on the file itself. Since no valid operations
require write permissions on the content files, mark them as read-only
so that the rest of the OS knows these files aren't meant to be written
to.

Resolves #95

Signed-off-by: Ryan Graham <r.m.graham@gmail.com>